fix staking wallet sidekick queries
